pyproject.toml
[tool.poetry]
name = "python-bdd"
version = "0.1.0"
description = "python avec bdd"
authors = ["Le concombre masqué"]
license = "MIT"
readme = "README.md"
packages = [{include = "python_bdd"}]

[tool.poetry.dependencies]
python = "^3.11"
docker = "^7.1.0"
textual = "^1.0.0"
mysql-connector-python = "^9.2.0"
faker = "^33.3.1"
mysql-connector = "^2.2.9"
flask = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
